Pregunta 16

+++ type.h
@@ -1,6 +1,6 @@
 #ifndef __TYPE_H__
 #define __TYPE_H__
 
-using RX=double;
+using RX= int;
 
 #endif
